reorganizing utilities
